Fixing some minor documentation stuff.
Also moved tests into separate files
